reshaping api/specs/webserver folder
