--- api/specs/webserver/openapi.yaml
@@ -1,114 +1,0 @@
-openapi: 3.0.0
-info:
-  title: "osparc-simcore web API"
-  version: 0.25.0
-  description: "API designed for the front-end app"
-  contact:
-    name: IT'IS Foundation
-    email: support@simcore.io
-  license:
-    name: MIT
-    url: https://github.com/ITISFoundation/osparc-simcore/blob/master/LICENSE
-servers:
-  - url: ""
-    description: 'Default server: requests directed to serving url'
-  - url: http://{host}:{port}
-    description: Development server
-    variables:
-      host:
-        default: "localhost"
-      port:
-        default: "8001"
-
-paths:
-
-
-  # PROJECT SERVICES -----------------------------------------------------------------
-
-  /v0/projects/{project_id}:open:
-    $ref: "./openapi-projects.yaml#/paths/~1projects~1{project_id}~1open"
-
-  /v0/projects/{project_id}/state:
-    $ref: "./openapi-projects.yaml#/paths/~1projects~1{project_id}~1state"
-
-  /v0/projects/{project_id}:xport: # do not change there "export" will not work
-    $ref: "./openapi-projects.yaml#/paths/~1projects~1{project_id}~1xport"
-
-  /v0/projects/{project_id}:close:
-    $ref: "./openapi-projects.yaml#/paths/~1projects~1{project_id}~1close"
-
-
-  /v0/projects/{project_id}/nodes/-/services:access:
-    $ref: "./openapi-projects-nodes.yaml#/paths/~1projects~1{project_id}~1nodes~1-~1services:access"
-
-  /v0/projects/{project_id}/nodes/-/preview:
-    $ref: "./openapi-projects-nodes.yaml#/paths/~1projects~1{project_id}~1nodes~1-~1preview"
-
-  /v0/projects/{project_id}/nodes/{node_id}/preview:
-    $ref: "./openapi-projects-nodes.yaml#/paths/~1projects~1{project_id}~1nodes~1{node_id}~1preview"
-
-  /v0/projects/{project_id}/metadata:
-    $ref: "./openapi-projects-metadata.yaml#/paths/~1projects~1{project_id}~1metadata"
-
-  /v0/projects/{project_uuid}/tags/{tag_id}:
-    $ref: "./openapi-projects.yaml#/paths/~1projects~1{project_uuid}~1tags~1{tag_id}"
-
-  # META-PROJECTS -------------------------------------------------------------------------
-
-  /v0/projects/{project_uuid}/checkpoint/{ref_id}/iterations:
-    $ref: "./openapi-meta-projects.yaml#/paths/~1projects~1{project_uuid}~1checkpoint~1{ref_id}~1iterations"
-
-  /v0/projects/{project_uuid}/checkpoint/{ref_id}/iterations/{iter_id}:
-    $ref: "./openapi-meta-projects.yaml#/paths/~1projects~1{project_uuid}~1checkpoint~1{ref_id}~1iterations~1{iter_id}"
-
-  /v0/projects/{project_uuid}/checkpoint/{ref_id}/iterations/-/results:
-    $ref: "./openapi-meta-projects.yaml#/paths/~1projects~1{project_uuid}~1checkpoint~1{ref_id}~1iterations~1-~1results"
-
-  /v0/projects/{project_uuid}/checkpoint/{ref_id}/iterations/{iter_id}/results:
-    $ref: "./openapi-meta-projects.yaml#/paths/~1projects~1{project_uuid}~1checkpoint~1{ref_id}~1iterations~1{iter_id}~1results"
-
-  # REPOSITORY -------------------------------------------------------------------------
-  /v0/repos/projects:
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects"
-
-  "/v0/repos/projects/{project_uuid}/checkpoints":
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects~1{project_uuid}~1checkpoints"
-
-  "/v0/repos/projects/{project_uuid}/checkpoints/HEAD":
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects~1{project_uuid}~1checkpoints~1HEAD"
-
-  "/v0/repos/projects/{project_uuid}/checkpoints/{ref_id}":
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects~1{project_uuid}~1checkpoints~1{ref_id}"
-
-  "/v0/repos/projects/{project_uuid}/checkpoints/{ref_id}:checkout":
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects~1{project_uuid}~1checkpoints~1{ref_id}:checkout"
-
-  "/v0/repos/projects/{project_uuid}/checkpoints/{ref_id}/workbench/view":
-    $ref: "./openapi-version-control.yaml#/paths/~1repos~1projects~1{project_uuid}~1checkpoints~1{ref_id}~1workbench~1view"
-
-  # TAGS -------------------------------------------------------------------------
-  /v0/tags:
-    $ref: "./openapi-tags.yaml#/paths/~1tags"
-
-  /v0/tags/{tag_id}:
-    $ref: "./openapi-tags.yaml#/paths/~1tags~1{tag_id}"
-
-  # RESOURCE USAGE ------------------------------------------------------------------
-
-  /v0/resource-usage/containers:
-    $ref: "./openapi-resource-usage.yaml#/paths/~1resource-usage~1containers"
-
-  # PUBLICATIONS -------------------------------------------------------------------------
-  /v0/publications/service-submission:
-    $ref: "./openapi-publications.yaml#/paths/~1publications~1service-submission"
-
-
-  # TASKS --------------------------------------------------------------------------
-  /v0/tasks:
-    $ref: "./openapi-tasks.yaml#/paths/~1tasks"
-
-  /v0/tasks/{task_id}:
-    $ref: "./openapi-tasks.yaml#/paths/~1tasks~1{task_id}"
-
-  /v0/tasks/{task_id}/result:
-    $ref: "./openapi-tasks.yaml#/paths/~1tasks~1{task_id}~1result"
